services/query-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
clickhouse-connect>=0.7.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
structlog>=24.1.0
prometheus-client>=0.19.0
httpx>=0.25.0
redis>=5.0.0
